DataRecorder/requirements.txt
cffi==1.15.0
cycler==0.11.0
fonttools==4.28.5
imutils==0.5.4
Jinja2==3.0.3
kiwisolver==1.3.2
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mplcursors==0.5.1
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pycparser==2.21
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.7.3
six==1.16.0
sounddevice==0.4.4
SoundFile==0.10.3.post1
